Fixes to format and structure
