+++ src/ctmain/ClayTargetPlugin.java
@@ -16,108 +16,39 @@ import org.bukkit.entity.Player;
 import org.bukkit.plugin.java.JavaPlugin;
 
 public class ClayTargetPlugin extends JavaPlugin {
-	public static boolean toggle = true;
-	public static List<Location> Locations = new ArrayList<Location>();
+	public static EventCommandListener listener;
 	@SuppressWarnings("unchecked")
 	@Override
 	public void onEnable() {
-		Bukkit.getPluginManager().registerEvents(new EventListener(this), this);
-if (getConfig() == null) {
-	this.saveDefaultConfig();
-}
-	if (getConfig().getList("Location") != null ) {
-	Locations = (List<Location>) getConfig().getList("Location");	
-	getLogger().info("Plugin has been Sucessfully Enabled");
-	}else {
-	getLogger().info("Plugin has been Sucessfully Enabled For the First Time");
-	getConfig().options().copyDefaults(true);
-	}
+		boolean toggle = true;
+		List<Location> Locations = new ArrayList<Location>();
+		if (getConfig() == null) {
+			this.saveDefaultConfig();
+		}
+		if (getConfig().getList("Location") != null ) {
+			Locations = (List<Location>) getConfig().getList("Location");	
+			getLogger().info("Plugin has been Sucessfully Enabled");
+		}else{
+			getLogger().info("Plugin has been Sucessfully Enabled For the First Time");
+			getConfig().options().copyDefaults(true);
+		}
+		listener = new EventCommandListener(this, Locations, toggle);
+		this.getServer().getPluginManager().registerEvents(listener, this);
+		this.getCommand("ctpregister").setExecutor(listener);
+		this.getCommand("ctpunregister").setExecutor(listener);
+		this.getCommand("ctpscore").setExecutor(listener);
+		this.getCommand("ctpscorereset").setExecutor(listener);
+		this.getCommand("ctptoggle").setExecutor(listener);
 	}
 	
 	@Override
 	public void onDisable() {
-		getConfig().set("Location", Locations);
+		getConfig().set("Location", listener.Locations);
 		saveConfig();
 		getLogger().info("Plugin has been Sucessfully Disabled");
 	}
 	
-	public boolean onCommand(CommandSender sender, Command cmd, String label, String[] args) {
-		if (cmd.getName().equalsIgnoreCase("ctpregister") && sender instanceof Player){
-			Player player = (Player) sender;
-			Block block = player.getTargetBlock((Set<Material>) null, 100);
-			if ( player.hasPermission("ctpperms")) {
-			if (block.getType().equals(Material.DISPENSER )) {
-				if(Locations.contains(block.getLocation())) {
-					sender.sendMessage("Already Registered");
-					return true;
-				}
-	Locations.add(block.getLocation());
-	sender.sendMessage("Dispenser Set. Use /ctpunregister to unregister it");
-				return true;
-			}else {sender.sendMessage("Block Is Not a Dispenser");}
-				return true;
-			}else {player.sendMessage("You do not have Permission to run this Commands");}
-				
-		
-		}
-		else if (cmd.getName().equalsIgnoreCase("ctpunregister") && sender instanceof Player) {
-			Player player = (Player) sender;
-			Block block = player.getTargetBlock((Set<Material>) null, 100);
-			if ( player.hasPermission("ctpperms")) {
-				if (Locations.contains(block.getLocation())) {
-			if (block.getType().equals(Material.DISPENSER )) {
-				Locations.remove(block.getLocation());
-				sender.sendMessage("Dispenser Unregistered");
-				return true;
-				}
-			else {sender.sendMessage("Block Is Not a Dispenser");}
-		 
-				return true;
-				}else {sender.sendMessage("This Block Isnt Registered");
-				return true;}
-			}else {player.sendMessage("You do not have Permission to run this Commands");}
-
-		}
-		else if (cmd.getName().equalsIgnoreCase("ctpscore") && sender instanceof Player) {
-			if ( EventListener.test.isEmpty()) {
-				Bukkit.broadcastMessage("Nobody Has Scored yet");
-				return true;	
-			}else {
-			for ( UUID key : EventListener.test.keySet() ) {
-				Player player = Bukkit.getPlayer(key);
-				sender.sendMessage(player.getDisplayName() + ": " + EventListener.test.get(key)) ;;
-			}
-			return true;
-			}
-		}
-			
-		else if (cmd.getName().equalsIgnoreCase("ctpscorereset") && sender instanceof Player) {
-			Player player = (Player) sender;
-			if ( player.hasPermission("ctpperms")) {
-		EventListener.test.clear();
-		sender.sendMessage("Scores Cleared");
-		
-		return true;
-		
-		}else {player.sendMessage("You do not have Permission to run this Commands");}
-		
-		}else if (cmd.getName().equalsIgnoreCase("ctptoggle") && sender instanceof Player) {
 	
-			
-			if (toggle == true) {
-				toggle = false;
-				sender.sendMessage("Toggle is now false");
-				return true;
-			}else if (toggle == false) { 
-				toggle = true; 
-				sender.sendMessage("Toggle is now True");
-				return true;
-				}
-			
-		}
-		
-		return false;
-		}
 	
 }
